Basic implementation of an inner join
This does not give any sugar to help prove the viability in the type
system, and the tests have manual implementations. Right now `Queriable`
needs to be explicitly implemented, since a generic impl like:

```rust
impl<A, B, SA, SB> Queriable<(SA, SB)> for (A, B) where
    A: Queriable<SA>,
    B: Queriable<SB>,
    SA: NativeSqlType,
    SB: NativeSqlType,
```

would overlap with the impl for tuples being queriable given `FromSql`.
Unsure how to resolve this, it seems like it might require either
specialization or negative type bounds.

This revealed an issue in the tuple implementation for `FromSql`. It
did work for nested tuples (and the type of this join is
`((i32, i32, String), (i32, String, Option<i16>))`). We can't just
assume every sub-type consumes exactly one column, and instead need to
ask it how many columns it consumed. I do not like this code structure,
but I do not have a better solution just yet.

Potential alternate solutions include having the type be flattened, or
using some sort of iterator instead of a row and index for
`FromSql::from_sql`.

The tuple implementation of `consumed_columns` has to do the + 0
nonsense because you can't actually expand a macro with a + character as
the middle. The `#[allow(unused_assignments)]` is because the generated
macro code actually increments idx one to many times (after reading the
last column, that number will never be read again)