Refactor the handling of tuples in `FromSql`
This hides the PG `Row` behind what is essentially an
`Iterator<Item=Any>`. This allows us to significantly improve the code
in the tuple impls of `FromSql`, and completely remove the
`consumed_columns()` nonsense.

With this change, we are assuming (and enforcing, really) that each
column is used only once in the returned data. That was already implied
in the way our types are structured, so this is fine.