Rip out the postgres crate, use libpq instead
I've had countless problems using this crate, and I believe it's time to
move away from it. Ultimately it's made several design decisions that
make it impossible to abstract over (in a way that will allow for
multiple DB backends)

One of the biggest examples is that Rows and Statement both take a
reference to their parent (Statement and Connection respectively), which
means that all uses of Rows must take place while Statement still lives,
and they cannot be moved together. All of this appears to be in order to
avoid heap allocation, which is silly since it's just representing heap
allocated data anyway.

Additionally, I can't find a good way to abstract over ToSql without
leaking that constraint up to the highest level. Ultimately what I'm
shooting for is a much more type safe API than they provide (requiring
the user to provide the types in a context where they cannot be
inferred)

Ultimately libpq is very easy to use given Rust's solid C FFI. While we
have to write a bit more boilerplate to have `ToSql` and `FromSql`
impls, I think this is ultimately going to be fine.

One thing that stands out to me at the moment are several places that
we're having to call `.unwrap()`. Ultimately everything that we're
asserting is an invariant which should always be true (numeric types
must hae the correct number of bytes, strings should always be UTF-8).
In the case where any of these were not true, it would not result in
memory safety. The only reason I can come up with that these would fail
to occur is if you're querying against the wrong type, which I think
panic is a perfectly acceptable answer there as we should statically
know the correct types. I will continue to explore whether we can at
least provide better errors in the case that they are broken, though.

I do, however, need to confirm the UTF-8 issue and decide what to do
there. I don't think having `ToSql` return a `Result` is going to work
here, since it should never fail, and it gets called during iteration,
not when we first perform the query.