Change `FromSql` to return an option
This change does not mean that invalid data won't panic. However, it
does allow us to avoid `unwrap`, and moves the control over the error
message up to a level where we might be able to get better debug info,
like the column name. It also opens the door for us to potentially offer
a panic-free option in the future.