Separate db -> primitives from primitives -> output
Basically the goal here is to separate out the iterator over the rows in
the result from the thing that performs the conversion, and introduced a
new trait to represent that. The reason for this is to be able to
ultimately implement a one to many type of associaiton.

If I have some representation of a type like `(A, Many<B>)`, then
`Many<B>` would implement the trait to deserialize multiple rows, but
would not implement the traits for a single row or a single column, much
like how tuples implement the trait for a full row, but not for a single
slice of bytes.

I wanted to use an iterator for this purpose, but I could not resolve
the lifetime constraints of the item, without tightly coupling to the
concrete implementation of `Row` used elsewhere.

There are still unresolved issues, like how I will deal with the
grouping, and how to deserialize starting from other columns, but these
will come later, and this should be a helpful foundation.